remove unused KspOptions and KSPCommandLineProcessor from analysis api
(cherry picked from commit 19c5b5528d2b2a26e49c22409beb277135996353)
